+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## XX.XX.X XXXX-XX-XX
+## 24.15.0 2025-06-02
 
 ### Connect
 * [CHANGED] Removed `PrivateBetaConnect` annotation and all usages. [Connect SDK](https://docs.stripe.com/connect/get-started-connect-embedded-components?platform=ios) is now GA.